requirements.txt
Django~=5.1.2
dj-database-url~=2.3.0
python-decouple~=3.8
djangorestframework~=3.15.2
django-filter~=24.3
psycopg~=3.2.3
black~=24.10.0
isort~=5.13.2
pre-commit~=4.0.1
autoflake~=2.3.1
drf-spectacular~=0.27.2
drf-spectacular-sidecar~=2024.7.1